backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.2.0
jinja2==3.1.3

# Notifications (optional - install manually if needed)
# sendgrid==6.11.0
# twilio==8.13.0
